Separated code out into multiple files and moved tests to their own directory.
