+++ backend/src/lib.rs
@@ -1,0 +1,5 @@
+
+pub mod types;
+pub mod cell;
+pub mod game;
+
