+++ backend/src/main.rs
@@ -1,11 +1,7 @@
 
-mod game;
-mod test;
-
-use crate::game::Twirl;
+use twirl::game::Twirl;
 
 fn main() {
-    let game = Twirl::new(9, 2);
-    println!("{game}");
+
 }
